Add support to seek in local mp4 files without SIDX boxes
Without this patch, the player relies on the presence of an `sidx` box to build up a seek map when playing an mp4 container.

This patch adds support for `mfra`, `tfra`, and `mfro` boxes and uses them to build the seek map instread. Unfortunately, the `mfra` box is located at the _end_ of the file. Reading this rather efficiently will therefore only work for local files.

The feature is also not implemented in a very clean way. We need to expose the underlying `DataSource` in the `ExtractorInput` and then perform some type checks to find the source URI. If the source is indeed a local file, we open a _new_ `DataSource` on that file and look for the `mfro` box at the end of the file and then use the offset exposed there to find and read the full `mfra` box and build up the seek map.

 In addition, the `moof` boxes might miss presentation time information, so we extended the parsers to use the generated (extended) seek map to find the presentation time for a given `moof` and its samples.